Added setting and code for vertex tax free tax code.

+++ src/InsiteCommerce.Web/Themes/Employee/Scripts/Custom/Cart/employee.review-and-pay.controller.ts
@@ -460,7 +460,7 @@
             errorLog.customerNumber = this.cart.billTo.customerNumber;
 
             var request = {
-                purchaseAmount: parseFloat(this.cart.orderGrandTotal.toFixed(2)),
+                purchaseAmount: parseInt(this.cart.orderGrandTotal.toFixed(2)),
                 purchaseCurrency: "840",
                 purchaseExponent: "2",
                 acctNumber: this.cart.paymentOptions.creditCard.cardNumber,
